Extract class ActiveStash::Index to its own file
