+++ lib/active_stash.rb
@@ -8,6 +8,7 @@ require "active_stash/schema_builder"
 require "active_stash/query_builder"
 require "active_stash/reflection"
 require "active_stash/relation"
+require "active_stash/index"
 require "active_stash/stash_indexes"
 require "active_stash/logger"
 require "active_stash/railtie" if defined?(Rails::Railtie)
